# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grooming Galore API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  category: string;
  price: number;
  /** @nullable */
  discountedPrice?: number | null;
  duration: string;
  isActive?: boolean;
  createdAt?: string;
}

export interface ServiceInput {
  name: string;
  description?: string;
  category: string;
  price: number;
  /** @nullable */
  discountedPrice?: number | null;
  duration: string;
  isActive?: boolean;
}

export interface ServiceUpdate {
  name?: string;
  description?: string;
  category?: string;
  price?: number;
  /** @nullable */
  discountedPrice?: number | null;
  duration?: string;
  isActive?: boolean;
}

export interface Offer {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  discountPercent: number | null;
  /** @nullable */
  discountAmount?: number | null;
  /** @nullable */
  validUntil?: string | null;
  isActive: boolean;
  createdAt?: string;
}

export interface OfferInput {
  title: string;
  description: string;
  /** @nullable */
  discountPercent?: number | null;
  /** @nullable */
  discountAmount?: number | null;
  /** @nullable */
  validUntil?: string | null;
  isActive: boolean;
}

export interface OfferUpdate {
  title?: string;
  description?: string;
  /** @nullable */
  discountPercent?: number | null;
  /** @nullable */
  discountAmount?: number | null;
  /** @nullable */
  validUntil?: string | null;
  isActive?: boolean;
}

export interface Testimonial {
  id: number;
  clientName: string;
  /** @nullable */
  role?: string | null;
  review: string;
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
  isVisible?: boolean;
  createdAt?: string;
}

export interface TestimonialInput {
  clientName: string;
  role?: string;
  review: string;
  rating: number;
  avatarUrl?: string;
  isVisible?: boolean;
}

export interface TestimonialUpdate {
  clientName?: string;
  role?: string;
  review?: string;
  rating?: number;
  avatarUrl?: string;
  isVisible?: boolean;
}

export interface Academy {
  id: number;
  name: string;
  description: string;
  courses?: string;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  fee?: number | null;
  /** @nullable */
  credentialInfo?: string | null;
  /** @nullable */
  internationalRecognition?: string | null;
  /** @nullable */
  inquiryEmail?: string | null;
  /** @nullable */
  inquiryPhone?: string | null;
  isActive?: boolean;
}

export interface AcademyUpdate {
  name?: string;
  description?: string;
  courses?: string;
  duration?: string;
  /** @nullable */
  fee?: number | null;
  credentialInfo?: string;
  internationalRecognition?: string;
  inquiryEmail?: string;
  inquiryPhone?: string;
  isActive?: boolean;
}

export interface Location {
  id: number;
  name: string;
  address: string;
  city?: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  openingHours?: string;
  /** @nullable */
  mapUrl?: string | null;
  isMain?: boolean;
}

export interface LocationInput {
  name: string;
  address: string;
  city?: string;
  phone: string;
  email?: string;
  openingHours: string;
  mapUrl?: string;
  isMain?: boolean;
}

export interface LocationUpdate {
  name?: string;
  address?: string;
  city?: string;
  phone?: string;
  email?: string;
  openingHours?: string;
  mapUrl?: string;
  isMain?: boolean;
}

export interface GalleryImage {
  id: number;
  url: string;
  alt: string;
  /** @nullable */
  category?: string | null;
  sortOrder?: number;
}

export interface GalleryImageInput {
  url: string;
  alt: string;
  category?: string;
  sortOrder?: number;
}

export interface Stats {
  totalServices: number;
  totalLocations: number;
  totalTestimonials: number;
  activeOffers: number;
  yearsOfExperience?: number;
  happyClients?: number;
  academyStudents?: number;
}

export type ListServicesParams = {
category?: string;
};