# lib/api-zod/src/generated/types/academy.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grooming Galore API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Academy {
  id: number;
  name: string;
  description: string;
  courses?: string;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  fee?: number | null;
  /** @nullable */
  credentialInfo?: string | null;
  /** @nullable */
  internationalRecognition?: string | null;
  /** @nullable */
  inquiryEmail?: string | null;
  /** @nullable */
  inquiryPhone?: string | null;
  isActive?: boolean;
}