# lib/api-zod/src/generated/types/location.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grooming Galore API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Location {
  id: number;
  name: string;
  address: string;
  city?: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  openingHours?: string;
  /** @nullable */
  mapUrl?: string | null;
  isMain?: boolean;
}