# lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Grooming Galore API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Offer {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  discountPercent: number | null;
  /** @nullable */
  discountAmount?: number | null;
  /** @nullable */
  validUntil?: string | null;
  isActive: boolean;
  createdAt?: string;
}